pyproject.toml
[tool.poetry]
name = "resene-colorchecker"
version = "0.1.0"
description = ""
authors = ["Nathan <thetruecoderman@gmail.com>"]

[tool.poetry.dependencies]
python = ">= 3.8, < 3.11"
colour-science = "^0.4.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
